+++ mipac/manager/note.py
@@ -6,6 +6,7 @@ from mipac.abstract.manager import AbstractManager
 from mipac.actions.note import ClientNoteActions, NoteActions
 from mipac.http import HTTPClient, Route
 from mipac.manager.favorite import FavoriteManager
+from mipac.manager.poll import PollManager
 from mipac.manager.reaction import ReactionManager
 
 if TYPE_CHECKING:
@@ -27,6 +28,9 @@ class ClientNoteManager(AbstractManager):
             note_id=note_id, session=session, client=client
         )
         self.favorite = FavoriteManager(
+            note_id=note_id, session=session, client=client
+        )
+        self.poll: PollManager = PollManager(
             note_id=note_id, session=session, client=client
         )
 
@@ -59,6 +63,9 @@ class NoteManager(AbstractManager):
             note_id=note_id, session=session, client=client
         )
         self._client: ClientNoteManager = ClientNoteManager(
+            note_id=note_id, session=session, client=client
+        )
+        self.poll: PollManager = PollManager(
             note_id=note_id, session=session, client=client
         )